+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### Features
+* Create and expose the Playground API, used for paper trading and backtesting [#71](https://github.com/Reiryoku-Technologies/Mida/pull/71)
 * Expose `withdrawCrypto()` interface [#73](https://github.com/Reiryoku-Technologies/Mida/pull/73)
 
 2022.4.0 - 16-08-2022